+++ neo/assistant/types.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/yaoapp/gou/rag/driver"
 	v8 "github.com/yaoapp/gou/runtime/v8"
+	api "github.com/yaoapp/yao/openai"
 )
 
 // API the assistant API interface
@@ -64,9 +65,9 @@ type Assistant struct {
 	Prompts     []Prompt                 `json:"prompts,omitempty"`     // AI Prompts
 	Flows       []map[string]interface{} `json:"flows,omitempty"`       // Assistant Flows
 	Script      *v8.Script               `json:"-" yaml:"-"`            // Assistant Script
-	API         API                      `json:"-" yaml:"-"`            // Assistant API
 	CreatedAt   int64                    `json:"created_at"`            // Creation timestamp
 	UpdatedAt   int64                    `json:"updated_at"`            // Last update timestamp
+	openai      *api.OpenAI              // OpenAI API
 }
 
 // File the file
