+++ neo/hooks.go
@@ -7,7 +7,6 @@ import (
 	"github.com/gin-gonic/gin"
 	jsoniter "github.com/json-iterator/go"
 	"github.com/yaoapp/gou/process"
-	"github.com/yaoapp/yao/neo/assistant"
 )
 
 // HookCreate create the assistant
@@ -67,51 +66,6 @@ func (neo *DSL) HookCreate(ctx Context, messages []map[string]interface{}, c *gi
 	}
 
 	return CreateResponse{AssistantID: assistantID, ChatID: ctx.ChatID}, nil
-}
-
-// HookAssistants query the assistant list from the assistant list hook
-func (neo *DSL) HookAssistants(ctx context.Context, param assistant.QueryParam) ([]assistant.Assistant, error) {
-	if neo.AssistantListHook == "" {
-		return nil, nil
-	}
-
-	// Create a context with 10 second timeout
-	timeoutCtx, cancel := context.WithTimeout(ctx, 10*time.Second)
-	defer cancel()
-
-	p, err := process.Of(neo.AssistantListHook, param)
-	if err != nil {
-		return nil, err
-	}
-
-	err = p.WithContext(timeoutCtx).Execute()
-	if err != nil {
-		return nil, err
-	}
-	defer p.Release()
-
-	// Check if context was canceled
-	if timeoutCtx.Err() != nil {
-		return nil, timeoutCtx.Err()
-	}
-
-	value := p.Value()
-	if value == nil {
-		return nil, nil
-	}
-
-	var list []assistant.Assistant
-	bytes, err := jsoniter.Marshal(value)
-	if err != nil {
-		return nil, err
-	}
-
-	err = jsoniter.Unmarshal(bytes, &list)
-	if err != nil {
-		return nil, err
-	}
-
-	return list, nil
 }
 
 // HookPrepare executes the prepare hook before AI is called
@@ -193,61 +147,4 @@ func (neo *DSL) HookWrite(ctx Context, messages []map[string]interface{}, respon
 	}
 
 	return result, nil
-}
-
-// HookMention query the mention list
-func (neo *DSL) HookMention(ctx context.Context, keywords string) ([]Mention, error) {
-
-	// Default Get the assistant list
-	if neo.MentionHook == "" {
-		var mentions []Mention
-		assistants := neo.GetAssistants()
-		for _, assistant := range assistants {
-			mentions = append(mentions, Mention{
-				ID:   assistant.ID,
-				Name: assistant.Name,
-				Type: "assistant",
-			})
-		}
-
-		return mentions, nil
-	}
-
-	// Create a context with 10 second timeout
-	timeoutCtx, cancel := context.WithTimeout(ctx, 10*time.Second)
-	defer cancel()
-
-	p, err := process.Of(neo.MentionHook, keywords)
-	if err != nil {
-		return nil, err
-	}
-
-	err = p.WithContext(timeoutCtx).Execute()
-	if err != nil {
-		return nil, err
-	}
-	defer p.Release()
-
-	// Check if context was canceled
-	if timeoutCtx.Err() != nil {
-		return nil, timeoutCtx.Err()
-	}
-
-	value := p.Value()
-	if value == nil {
-		return nil, nil
-	}
-
-	var list []Mention
-	bytes, err := jsoniter.Marshal(value)
-	if err != nil {
-		return nil, err
-	}
-
-	err = jsoniter.Unmarshal(bytes, &list)
-	if err != nil {
-		return nil, err
-	}
-
-	return list, nil
 }
